Updated to compatible to MacOS Mojave under anaconda python 2.7 environment

+++ python/caffe/detector.py
@@ -137,6 +137,7 @@ class Detector(caffe.Net):
         crop: cropped window.
         """
         # Crop window from the image.
+        window=window.astype(np.int64)
         crop = im[window[0]:window[2], window[1]:window[3]]
 
         if self.context_pad:
@@ -171,10 +172,15 @@ class Detector(caffe.Net):
 
             # collect with context padding and place in input
             # with mean padding
+            box=box.astype(np.int64)
             context_crop = im[box[0]:box[2], box[1]:box[3]]
             context_crop = caffe.io.resize_image(context_crop, (crop_h, crop_w))
             crop = np.ones(self.crop_dims, dtype=np.float32) * self.crop_mean
-            crop[pad_y:(pad_y + crop_h), pad_x:(pad_x + crop_w)] = context_crop
+            crop[int(pad_y):int(pad_y + crop_h), int(pad_x):int(pad_x + crop_w)] = context_crop
+            # context_crop = im[box[0]:box[2], box[1]:box[3]]
+            # context_crop = caffe.io.resize_image(context_crop, (crop_h, crop_w))
+            # crop = np.ones(self.crop_dims, dtype=np.float32) * self.crop_mean
+            # crop[pad_y:(pad_y + crop_h), pad_x:(pad_x + crop_w)] = context_crop
 
         return crop
